+++ src/main/java/com/uhcchampions/uhc/Commands/GMC.java
@@ -1,4 +1,4 @@
-package com.uhcchampions.uhc;
+package com.uhcchampions.uhc.Commands;
 
 import org.bukkit.ChatColor;
 import org.bukkit.GameMode;
