+++ src/main/java/com/uhcchampions/uhc/Commands/PvPCommand.java
@@ -1,32 +1,35 @@
-package com.uhcchampions.uhc;
+package com.uhcchampions.uhc.Commands;
 
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
-import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
-import org.bukkit.event.entity.EntityDamageByEntityEvent;
 
-public class PvPCommand2 implements CommandExecutor, Listener {
+public class PvPCommand implements CommandExecutor, Listener {
+
+
     @Override
     public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
         if(sender instanceof Player) {
             Player player = (Player) sender;
 
-            if(player.hasPermission("pvpe.use")) {
-                player.sendMessage(ChatColor.GREEN + "PVP ENABLED!");
-                player.getLocation().getWorld().setPVP(true);
+            if(player.hasPermission("pvp.use")) {
+
+                player.sendMessage(ChatColor.RED + "PVP DISABLED!");
+                player.getLocation().getWorld().setPVP(false);
 
 
             } else {
-                player.sendMessage(ChatColor.RED + "No permission!");
-            }
+            player.sendMessage(ChatColor.RED + "No permission!");
+        }
+
         }
 
 
 
         return false;
     }
-}
+
+    }
