+++ src/main/java/com/uhcchampions/uhc/Lag.java
@@ -5,7 +5,6 @@ public class Lag
 {
     public static int TICK_COUNT= 0;
     public static long[] TICKS= new long[600];
-    public static long LAST_TICK= 0L;
 
     public static double getTPS()
     {
@@ -21,16 +20,6 @@ public class Lag
         long elapsed = System.currentTimeMillis() - TICKS[target];
 
         return ticks / (elapsed / 1000.0D);
-    }
-
-    public static long getElapsed(int tickID)
-    {
-        if (TICK_COUNT- tickID >= TICKS.length)
-        {
-        }
-
-        long time = TICKS[(tickID % TICKS.length)];
-        return System.currentTimeMillis() - time;
     }
 
     public void run()
