+++ src/main/java/com/uhcchampions/uhc/Listeners/pceDeath.java
@@ -1,19 +1,14 @@
-package com.uhcchampions.uhc;
+package com.uhcchampions.uhc.Listeners;
 
 import org.bukkit.*;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
-import org.bukkit.block.BlockState;
 import org.bukkit.block.Skull;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.PlayerDeathEvent;
-import org.bukkit.inventory.ItemStack;
-import org.bukkit.inventory.meta.SkullMeta;
-
-import java.util.UUID;
 
 public class pceDeath implements Listener {
